feat: added the sitemap.js utility and used its features for adding and editing partners
