+++ src/routes/api/addPartner/+server.js
@@ -1,163 +1,16 @@
 import { gql } from 'graphql-request';
 import { hygraph } from '$lib/utils/hygraph.js';
 import getQueryAddPartner from '$lib/queries/addPartner';
-import getQueryAddUrl from '$lib/queries/addUrl';
-import getQueryUrl from '$lib/queries/url';
 import getQueryUpdatePartnerUrls from '$lib/queries/updateUrlsPartner';
-import Sitemapper from 'sitemapper';
-import axios from 'axios';
-import { parseHTML } from 'linkedom';
+import {
+  formatUrl,
+  getSitemapPromises,
+  pickFirstSitemap,
+  crawlUrls,
+  processUrls,
+  delay
+} from '$lib/utils/sitemap.js';
 
-// Utility function to validate URLs
-function isValidUrl(url) {
-  return !url.includes('/document') && !url.includes('/documents');
-}
-
-// Delay helper
-function delay(ms) {
-  return new Promise(resolve => setTimeout(resolve, ms));
-}
-
-// Ensure URL ends with '/'
-async function formatUrl(rawUrl, sendUpdate) {
-  const url = rawUrl.endsWith('/') ? rawUrl : rawUrl + '/';
-  await sendUpdate({ status: 'Url verwerken', type: 'done' });
-  await delay(1000);
-  return url;
-}
-
-// Attempt to fetch sitemap URLs
-function getSitemapPromises(baseUrl, sitemapPaths, sendUpdate) {
-  return sitemapPaths.map(path =>
-    (async () => {
-      try {
-        await sendUpdate({ status: `Sitemap checken in pad ${path}`, type: 'done'  });
-        await delay(1000);
-        const siteMap = new Sitemapper({ url: baseUrl + path, timeout: 15000 });
-        const { sites } = await siteMap.fetch();
-        return (sites || []).filter(isValidUrl);
-      } catch (err) {
-        await sendUpdate({ status: `Sitemap niet gevonden: ${path}`, type: 'error' });
-        await delay(1000);
-        return [];
-      }
-    })()
-  );
-}
-
-// Pick the first non-empty sitemap result
-async function pickFirstSitemap(promises, sendUpdate) {
-  const results = await Promise.all(promises);
-  for (const group of results) {
-    if (group.length > 0) {
-      await sendUpdate({ status: 'Sitemap gevonden', type: 'done' });
-      await delay(1000);
-      return group;
-    }
-  }
-  return [];
-}
-
-// Crawl pages to extract links
-async function crawlUrls(startUrl, sendUpdate) {
-  await sendUpdate({ status: 'Sitemap niet gevonden, urls worden handmatig opgehaald', type: 'warning'  });
-  await delay(1000);
-  const visited = new Set();
-  const toVisit = [startUrl];
-
-  async function getLinksFromPage(pageUrl) {
-    const res = await axios.get(pageUrl);
-    const { document } = parseHTML(res.data);
-    const links = Array.from(document.querySelectorAll('a'))
-      .map(a => a.getAttribute('href'))
-      .filter(href => href && !href.startsWith('#'))
-      .map(href => new URL(href, pageUrl).href)
-      .filter(h => h.startsWith(startUrl))
-      .filter(isValidUrl);
-    return [...new Set(links)];
-  }
-
-  while (toVisit.length) {
-    const current = toVisit.shift();
-    if (visited.has(current)) continue;
-    visited.add(current);
-    await sendUpdate({ status: `Bezoek: ${current}`, type: 'done'  });
-    await delay(1000);
-    try {
-      const found = await getLinksFromPage(current);
-      for (const link of found) {
-        if (!visited.has(link) && !toVisit.includes(link)) {
-          toVisit.push(link);
-        }
-      }
-    } catch (err) {
-      await sendUpdate({ status: `Error: ${current}: ${err.message}`, type: 'error' });
-      await delay(1000);
-    }
-  }
-
-  await sendUpdate({ status: 'Gevonden urls', count: visited.size, type: 'done' });
-  await delay(1000);
-  return Array.from(visited);
-}
-
-// Add or skip partner
-async function addPartner({ id, name, url, slug, count }, sendUpdate) {
-  if (!id) {
-    const query = getQueryAddPartner(gql, name, url, slug, count);
-    await hygraph.request(query);
-    await sendUpdate({ status: 'Partner toegevoegd', type: 'done' });
-    await delay(1000);
-  } else {
-    await sendUpdate({ status: 'Partner bestaat al', type: 'warning' });
-    await delay(1000);
-  }
-}
-
-// Process individual URLs
-async function processUrls(urls, slug, sendUpdate) {
-  let total = urls.length;
-  const failed = {};
-  await sendUpdate({ status: 'Urls worden toegevoegd aan de database', type: 'done' });
-  await delay(1000);
-
-  for (let i = 1; i < urls.length; i++) {
-    const link = urls[i];
-    const path = new URL(link).pathname;
-    let urlSlug = (slug + path).replace(/\//g, '-');
-    await sendUpdate({ status: `Verwerk URL ${i + 1}/${urls.length}`, type: 'done' });
-    await delay(1000);
-
-    try {
-      const checkQuery = getQueryUrl(gql, urlSlug);
-      const checkRes = await hygraph.request(checkQuery);
-      if (checkRes.url) {
-        await sendUpdate({ status: `Url bestaat al: ${checkRes.url.slug}`, type: 'warning' });
-        total--;
-      } else {
-        await sendUpdate({ status: `Voeg toe: ${link}`, type: 'done' });
-        const addQuery = getQueryAddUrl(gql, urlSlug, link, slug, path);
-        await hygraph.request(addQuery);
-      }
-    } catch (err) {
-      failed[link] = err.message;
-    }
-
-    await delay(150);
-  }
-
-  return { total, failed };
-}
-
-// Update partner with final URL count
-async function updatePartnerUrls(slug, count, sendUpdate) {
-  const query = getQueryUpdatePartnerUrls(gql, slug, count);
-  await hygraph.request(query);
-  await sendUpdate({ status: 'Partner bijgewerkt', type: 'done' });
-  await delay(1000);
-}
-
-// Main handler
 export async function POST({ request }) {
   const form = await request.formData();
   const name = form.get('name');
@@ -185,7 +38,8 @@ export async function POST({ request }) {
           closed = true;
         }
       };
-      const sendUpdate = async msg => controller.enqueue(enc.encode(`data: ${JSON.stringify(msg)}\n\n`));
+      const sendUpdate = async msg =>
+        controller.enqueue(enc.encode(`data: ${JSON.stringify(msg)}\n\n`));
 
       (async () => {
         try {
@@ -195,7 +49,6 @@ export async function POST({ request }) {
           let urls = [];
           if (toggle) {
             url = await formatUrl(rawUrl, sendUpdate);
-
             const promises = getSitemapPromises(url, sitemapPaths, sendUpdate);
             const sitemapUrls = await pickFirstSitemap(promises, sendUpdate);
 
@@ -204,13 +57,25 @@ export async function POST({ request }) {
               : await crawlUrls(url, sendUpdate);
           }
 
-          await addPartner({ id, name, url, slug, count: urls.length }, sendUpdate);
+          // Add or update partner
+          if (!id) {
+            const query = getQueryAddPartner(gql, name, url, slug, urls.length);
+            await hygraph.request(query);
+            await sendUpdate({ status: 'Partner toegevoegd', type: 'done' });
+            await delay(1000);
+          } else {
+            await sendUpdate({ status: 'Partner bestaat al', type: 'warning' });
+            await delay(1000);
+          }
 
+          // Process URLs if toggle is on
           if (toggle && urls.length) {
             const { total } = await processUrls(urls, slug, sendUpdate);
-            await updatePartnerUrls(slug, total, sendUpdate);
-            await sendUpdate({ status: 'Alle urls zijn toegevoegd', type: 'done' });
+            const updateQuery = getQueryUpdatePartnerUrls(gql, slug, total);
+            await hygraph.request(updateQuery);
+            await sendUpdate({ status: 'Partner bijgewerkt', type: 'done' });
             await delay(1000);
+            await sendUpdate({ status: 'Alle urls zijn toegevoegd', type: 'done' });
           }
         } catch (err) {
           await sendUpdate({ status: err.message, type: 'error' });
